+++ requirements.txt
@@ -1,0 +1,15 @@
+blinker==1.9.0
+click==8.2.1
+colorama==0.4.6
+Flask==3.1.1
+Flask-SQLAlchemy==3.1.1
+Flask-WTF==1.2.2
+greenlet==3.2.3
+itsdangerous==2.2.0
+Jinja2==3.1.6
+MarkupSafe==3.0.2
+SQLAlchemy==2.0.42
+typing_extensions==4.14.1
+Werkzeug==3.1.3
+WTForms==3.2.1
+
