+++ static/maps.css
@@ -1,0 +1,110 @@
+body {
+    font-family: 'Roboto', 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    margin: 0;
+    padding: 0;
+    min-height: 100vh;
+    background-color: #f4f0fa;
+}
+
+.nav {
+    background-color: #4b0082;
+    padding: 1rem 2rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+}
+
+.nav-brand {
+    color: white;
+    font-size: 1.5rem;
+    font-weight: bold;
+}
+
+.nav-links {
+    list-style: none;
+    display: flex;
+    gap: 1.5rem;
+}
+
+.nav-links li a {
+    color: white;
+    text-decoration: none;
+    font-size: 1rem;
+    transition: color 0.3s;
+}
+
+.nav-links li a:hover, .nav-links li a.active {
+    color: #e6e6fa;
+}
+
+.container {
+    max-width: 1200px;
+    width: 100%;
+    padding: 2rem;
+    margin: 2rem auto;
+    background: white;
+    box-shadow: 0 4px 15px rgba(75, 0, 130, 0.2);
+    border-radius: 10px;
+}
+
+h1 {
+    text-align: center;
+    margin-bottom: 1.5rem;
+    color: #4b0082;
+}
+
+.form-group {
+    margin-bottom: 1.2rem;
+}
+
+.form-control {
+    width: 100%;
+    padding: 0.7rem;
+    border: 1px solid #ccccff;
+    border-radius: 5px;
+    font-size: 1rem;
+    background-color: #fafaff;
+}
+
+.form-control:focus {
+    outline: none;
+    border-color: #9900cc;
+    box-shadow: 0 0 5px rgba(153, 0, 204, 0.4);
+}
+
+.form-control-label {
+    display: block;
+    margin-bottom: 0.5rem;
+    font-weight: bold;
+    color: #4b0082;
+}
+
+.btn {
+    background-color: transparent;
+    color: #9900cc;
+    padding: 0.7rem 1.5rem;
+    border: 2px solid #9900cc;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 1rem;
+    transition: 0.3s ease-in-out;
+}
+
+.btn:hover {
+    background-color: #9900cc;
+    color: white;
+}
+
+#map {
+    height: 500px;
+    width: 100%;
+    margin-top: 1.5rem;
+    border-radius: 8px;
+}
+
+#travel-info {
+    margin-top: 1.5rem;
+    font-size: 1rem;
+    color: #4b0082;
+}
