+++ templates/dashboard.html
@@ -2,95 +2,63 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Dashboard Page</title>
-    <link rel="stylesheet" href="static/style2.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Dashboard - SendIT</title>
+    <link rel="stylesheet" href="/static/style1.css">
 </head>
 <body>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>SignIT | Dashboard</title>
-    <link rel="stylesheet" href="dsh.css">
-<body>
-    <div class="dashboard-container">
-       <div class="top-part">
-              <ul class="nav-list">
-                  <li><a href="/">Home</a></li>
-                  <li><a href="/dashboard">Dashboard</a></li>
-                  <li><a href="/profile">Profile</a></li>
-                  <li><a href="/logout">Logout</a></li>
-               </ul>
-       </div>
-
-        <main class="main-content">
-            <h1>Welcome User</h1>
-            <section class="create-parcel">
-                <h2>Create New Parcel</h2>
-                <form class="create-parcel form" action="{{ url_for('dashboard_page')}}" method="POST">
-
-                      {{form.hidden_tag()}}
-                <div class="form-group">
-                      {{form.pickup_location.label(class = "form-control-label")}}
-                      {{form.pickup_location(class = "form-control form-control")}}
-
-                </div>
-
-                <div class="form-group">
-                       {{form.destination.label(class = "form-control-label")}}
-                       {{form.destination(class = "form-control form-control")}}
-
-                  </div>
-                  <div class="form-group">
-                       {{form.weight.label(class = "form-control-label")}}
-                       {{form.weight(class = "form-control form-control")}}
-
-                   </div>
-                   <div class="form-group">
-                       {{form.submit(class = "btn btn-outline-info")}}
-
-
-                    </div>
-                </form>
-
-
-            </section>
-
-            <section class="parcel-list">
-                <h2>Your Parcels</h2>
-                <table>
-                    <thead>
-                        <tr>
-                            <th>ID</th>
-                            <th>Pickup</th>
-                            <th>Destination</th>
-                            <th>Status</th>
-                            <th>Action</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <tr>
-                            <td>#1234</td>
-                            <td>Abuja</td>
-                            <td>Lagos</td>
-                            <td>In Transit</td>
-                            <td>
-                              <div class="button-group">
-                                  <button type="submit" class="btn btn-change">Change Destination</button>
-                                  <button type="submit" class="btn btn-cancel">Cancel</button>
-                              </div>
-                            </td>
-                        </tr>
-
-                    </tbody>
-                </table>
-            </section>
-        </main>
+    <nav class="nav">
+        <div class="nav-brand">SendIT</div>
+        <ul class="nav-links">
+            <li><a href="/">Home</a></li>
+            <li><a href="/dashboard" class="active">Dashboard</a></li>
+            <li><a href="/profile">Profile</a></li>
+            <li><a href="/logout">Logout</a></li>
+        </ul>
+    </nav>
+    <div class="container">
+        <h1>Welcome User</h1>
+        <h2>Create New Parcel</h2>
+        <form method="POST">
+            {{ form.hidden_tag() }}
+            <div class="form-group">
+                {{ form.pickup_location.label(class="form-control-label") }}
+                {{ form.pickup_location(class="form-control") }}
+            </div>
+            <div class="form-group">
+                {{ form.destination.label(class="form-control-label") }}
+                {{ form.destination(class="form-control") }}
+            </div>
+            <div class="form-group">
+                {{ form.weight.label(class="form-control-label") }}
+                {{ form.weight(class="form-control") }}
+            </div>
+            {{ form.submit(class="btn btn-outline-info") }}
+        </form>
+        <h2>Your Parcels</h2>
+        <table class="parcel-table">
+            <thead>
+                <tr>
+                    <th>ID</th>
+                    <th>Pickup</th>
+                    <th>Destination</th>
+                    <th>Status</th>
+                    <th>Action</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <td>#1234</td>
+                    <td>Abuja</td>
+                    <td>Lagos</td>
+                    <td>In Transit</td>
+                    <td>
+                        <a href="#" class="btn btn-small btn-outline-info">Change Destination</a>
+                        <a href="#" class="btn btn-small btn-outline-danger">Cancel</a>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
     </div>
-
-</body>
-</html>
-
 </body>
 </html>
