+++ templates/login.html
@@ -2,34 +2,28 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Login Page</title>
-    <link rel="stylesheet" href="static/style1.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Login - SendIT</title>
+    <link rel="stylesheet" href="/static/style1.css">
 </head>
 <body>
-<div class="container ">
-    <form method="POST" action="{{ url_for('login_page') }}">
-        {{form.hidden_tag()}}
-        <div class="form-group">
-            {{form.email.label(class = "form-control-label")}}
-            {{form.email(class = "form-control form-control")}}
-
-        </div>
-        <div class="form-group">
-            {{form.password.label(class = "form-control-label")}}
-            {{form.password(class = "form-control form-control")}}
-
-        </div>
-        <div class="form-group">
-            {{form.submit(class = "btn btn-outline-info")}}
-
-
-        </div>
+    <div class="container">
+        <h1>Login</h1>
+        <form method="POST">
+            {{ form.hidden_tag() }}
+            <div class="form-group">
+                {{ form.email.label(class="form-control-label") }}
+                {{ form.email(class="form-control") }}
+            </div>
+            <div class="form-group">
+                {{ form.password.label(class="form-control-label") }}
+                {{ form.password(class="form-control") }}
+            </div>
+            {{ form.submit(class="btn btn-outline-info") }}
         </form>
-    <div class="border-top pt-3">
-        <small class="text muted"></small>
-        <p>You don't have an account?<a href="{{url_for('register_page')}}">SignUp</a> </p>
+        <div class="border-top">
+            <p>You don't have an account? <a href="/register">SignUp</a></p>
+        </div>
     </div>
-</div>
-
 </body>
 </html>
