+++ templates/profile.html
@@ -2,58 +2,53 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Profile Page</title>
-    <link rel="stylesheet" href="static/style2.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>User Profile - SendIT</title>
+    <link rel="stylesheet" href="/static/style1.css">
 </head>
 <body>
-       <div class="top-part">
-              <ul class="nav-list">
-                  <li><a href="/">Home</a></li>
-                  <li><a href="/dashboard">Dashboard</a></li>
-                  <li><a href="/profile">Profile</a></li>
-                  <li><a href="/logout">Logout</a></li>
-               </ul>
-       </div>
-
-       <main class="main-content">
-           <h1>User Profile</h1>
-
-           <section class="stats">
-              <div class="stat-box">
-                  <h3>Delivered Parcels</h3>
-                  <p>3</p>
-              </div>
-              <div class="stat-box">
-                  <h3>In Transit</h3>
-                  <p>2</p>
-              </div>
-           </section>
-
-           <section class="parcel-list">
-                <h2>All Parcels</h2>
-                <table>
-                   <thead>
-                        <tr>
-                             <th>ID</th>
-                             <th>Pickup</th>
-                             <th>Destination</th>
-                             <th>Status</th>
-                             <th>Created At</th>
-                        </tr>
-                   </thead>
-                   <tbody>
-                         <tr>
-                             <td>#001</td>
-                             <td>Lagos</td>
-                             <td>Ibadan</td>
-                             <td>Delivered</td>
-                             <td>2025-07-20</td>
-                          </tr>
-
-                   </tbody>
-                 </table>
-           </section>
-       </main>
-
+    <nav class="nav">
+        <div class="nav-brand">SendIT</div>
+        <ul class="nav-links">
+            <li><a href="/">Home</a></li>
+            <li><a href="/dashboard">Dashboard</a></li>
+            <li><a href="/profile" class="active">Profile</a></li>
+            <li><a href="/logout">Logout</a></li>
+        </ul>
+    </nav>
+    <div class="container">
+        <h1 class="profile-header">User Profile</h1>
+        <div class="profile-stats">
+            <div class="stat-card">
+                <h3>Delivered Parcels</h3>
+                <p class="stat-number">3</p>
+            </div>
+            <div class="stat-card">
+                <h3>In Transit</h3>
+                <p class="stat-number">2</p>
+            </div>
+        </div>
+        <h2>All Parcels</h2>
+        <table class="parcel-table">
+            <thead>
+                <tr>
+                    <th>ID</th>
+                    <th>Pickup</th>
+                    <th>Destination</th>
+                    <th>Status</th>
+                    <th>Created At</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <td>#001</td>
+                    <td>Lagos</td>
+                    <td>Ibadan</td>
+                    <td>Delivered</td>
+                    <td>2025-07-20</td>
+                </tr>
+            </tbody>
+        </table>
+    </div>
 </body>
 </html>
